Organize code by splitting large Program into separate classes
